spaceapps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
catboost==1.2.8
pytorch-tabnet==4.1.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
python-multipart==0.0.6
google-genai==0.2.2
python-dotenv==1.0.0
